@@ -1,14 +1,14 @@
-﻿using Bookinist.Service.UserDialogService;
+﻿using Bookinist.Service.Interfaces;
+using Bookinist.Service.UserDialogService;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Bookinist.Service
+namespace Bookinist.Service;
+
+public static class ServiceRegistration
 {
-    public static class ServiceRegistration
-    {
-        public static IServiceCollection RegisterServices(this IServiceCollection services)
-        {
-            services.AddTransient<IUserDialogService, WindowsUserDialogService>();
-            return services;
-        }
-    }
+    public static IServiceCollection RegisterServices(this IServiceCollection services) => services
+        .AddTransient<IUserDialogService, WindowsUserDialogService>()
+        .AddSingleton<ISalesService, SalesService>()
+        ;
+        
 }
