@@ -1,14 +1,12 @@
 ﻿using Bookinist.ViewModels.MainWindowVm;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Bookinist.ViewModels
+namespace Bookinist.ViewModels;
+
+public static class ViewModelRegistration
 {
-    public static class ViewModelRegistration
-    {
-        public static IServiceCollection RegisterViewModels(this IServiceCollection services)
-        {
-            services.AddSingleton<MainWindowViewModel>();
-            return services;
-        }
-    }
+    public static IServiceCollection RegisterViewModels(this IServiceCollection services) => services
+        .AddSingleton<MainWindowViewModel>()
+        ;
+        
 }
